requirements.txt
torch==2.0.1
torchvision==0.15.2
accelerate==0.27.2
fair-esm
easydict
pyyaml
lmdb
scikit-learn
pandas
decorator
ipython
networkx
tqdm
matplotlib
deepspeed==0.14.0
datasets
peft==0.9.0
transformers==4.39.0
numpy<2.0
wandb
hydra-core
sentencepiece
h5py
bert_score
nltk
rouge_score
git+https://github.com/microsoft/tutel@main